requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.2.0
awkward==1.8.0
black==23.3.0
brotlipy==0.7.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
colorlover==0.3.0
cryptography==39.0.1
Cython==0.29.33
dash==2.5.1
dash-bootstrap-components==1.2.0
dash-cytoscape==0.3.0
datasets==2.10.1
dill==0.3.6
filelock==3.9.0
flake8==6.0.0
frozenlist==1.3.3
fsspec==2023.1.0
hdbscan==0.8.33
huggingface-hub==0.12.1
idna==3.4
imageio==2.26.0
importlib-metadata==6.0.0
joblib==1.2.0
llvmlite==0.39.1
matplotlib==3.5.2
multidict==6.0.4
multiprocess==0.70.14
natsort==8.1.0
networkx==2.8.2
numba==0.56.4
numpy==1.23.5
opencv-contrib-python-headless==4.6.0.66
packaging==23.0
pandas==1.5.3
pip==23.0.1
platformdirs==3.0.0
pooch==1.7.0
pre-commit==3.2.2
pyarrow==11.0.0
pycparser==2.21
pynndescent==0.5.8
pyOpenSSL==23.0.0
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2022.7.1
PyYAML==6.0
regex==2022.10.31
requests==2.28.2
responses==0.18.0
rouge==1.0.0
ruamel.yaml==0.17.20
ruamel.yaml.clib==0.2.6
sacremoses==0.0.53
scikit-learn==1.2.1
scipy==1.10.1
seaborn==0.12.2
setuptools==67.4.0
six==1.16.0
threadpoolctl==3.1.0
tokenizers==0.13.2
torch==1.13.1
tqdm==4.64.1
transformers==4.26.1
typing_extensions==4.4.0
umap-learn==0.5.3
urllib3==1.26.14
wheel==0.38.4
xxhash==3.2.0
yarl==1.8.2
zipp==3.15.0
